Clean code + update subclass
